+++ src/main/java/io/fabric8/maven/docker/assembly/DockerAssemblyManager.java
@@ -186,22 +186,36 @@ public class DockerAssemblyManager {
         }
     }
 
-    private void verifyGivenDockerfile(File dockerFile, BuildImageConfiguration buildConfig, FixedStringSearchInterpolator interpolator, Logger log) throws IOException {
+    // visible for testing
+    void verifyGivenDockerfile(File dockerFile, BuildImageConfiguration buildConfig, FixedStringSearchInterpolator interpolator, Logger log) throws IOException {
         AssemblyConfiguration assemblyConfig = buildConfig.getAssemblyConfiguration();
-        if (assemblyConfig != null) {
-            String name = assemblyConfig.getName();
+        if (assemblyConfig == null) {
+            return;
+        }
+
+        String name = assemblyConfig.getName();
             for (String keyword : new String[] { "ADD", "COPY" }) {
                 List<String[]> lines = DockerFileUtil.extractLines(dockerFile, keyword, interpolator);
                 for (String[] line : lines) {
-                    // contains an ADD/COPY ... targetDir .... All good.
-                    if (!line[0].startsWith("#") && line.length > 1 && line[1].contains(name)) {
-                        return;
+                    if (!line[0].startsWith("#")) {
+                        // Skip command flags like --chown
+                        int i;
+                        for (i = 1; i < line.length; i++) {
+                            String component = line[i];
+                            if (!component.startsWith("--")) {
+                                break;
+                            }
+                        }
+
+                        // contains an ADD/COPY ... targetDir .... All good.
+                        if (i < line.length && line[i].contains(name)) {
+                            return;
+                        }
                     }
                 }
             }
-            log.warn("Dockerfile %s does not contain an ADD or COPY directive to include assembly created at %s. Ignoring assembly.",
-                     dockerFile.getPath(), name);
-        }
+        log.warn("Dockerfile %s does not contain an ADD or COPY directive to include assembly created at %s. Ignoring assembly.",
+                 dockerFile.getPath(), name);
     }
 
     /**