+++ doc/changelog.md
@@ -2,7 +2,8 @@
 
 * **0.22-SNAPSHOT**
   - Support relative paths when binding volumes in `docker-compose.yml` (#846)
-  - Allo the session token for AWS authetication to be oncluded in order to allow temporary security credentials provided by the AWS Security Token Service (AWS STS) to sign requests (#883)
+  - Allow  the session token for AWS authetication to be oncluded in order to allow temporary security credentials provided by the AWS Security Token Service (AWS STS) to sign requests (#883)
+  - Add support for credential helper to authenticate against a registry (#821)
 
 * **0.22.1** (2017-08-28)
   - Allow Docker compose version "2", too ([#829](https://github.com/fabric8io/docker-maven-plugin/issues/829))